refactor: better format separation
