+++ src/gaussian/formats/planar_4d_hierarchy.rs
@@ -1,5 +1,5 @@
 // TODO: gaussian cloud 4d with temporal hierarchy
-use crate::gaussian::packed::PlanarGaussian4dHandle;
+use crate::gaussian::formats::planar_4d::PlanarGaussian4dHandle;
 
 
 pub struct TemporalGaussianLevel {
