+++ src/query/select.rs
@@ -4,7 +4,10 @@ use bevy_interleave::prelude::*;
 use crate::{
     gaussian::{
         interface::CommonCloud,
-        packed::{Gaussian3d, Gaussian4d},
+        formats::{
+            planar_3d::Gaussian3d,
+            planar_4d::Gaussian4d,
+        },
     }, io::codec::CloudCodec,
 };
