+++ src/sort/mod.rs
@@ -35,11 +35,11 @@ use crate::{
     camera::GaussianCamera,
     CloudSettings,
     gaussian::{
-        interface::CommonCloud,
-        packed::{
-            Gaussian3d,
-            Gaussian4d,
+        formats::{
+            planar_3d::Gaussian3d,
+            planar_4d::Gaussian4d,
         },
+        interface::CommonCloud,
     },
 };
