+++ src/sort/std.rs
@@ -9,8 +9,11 @@ use crate::{
     camera::GaussianCamera,
     CloudSettings,
     gaussian::{
+        formats::{
+            planar_3d::Gaussian3d,
+            planar_4d::Gaussian4d,
+        },
         interface::CommonCloud,
-        packed::Gaussian3d,
     },
     sort::{
         SortConfig,
@@ -28,6 +31,7 @@ pub struct StdSortPlugin;
 impl Plugin for StdSortPlugin {
     fn build(&self, app: &mut App) {
         app.add_systems(Update, std_sort::<Gaussian3d>);
+        app.add_systems(Update, std_sort::<Gaussian4d>);
     }
 }
